third_party/blink/web_tests/external/wpt/editing/other/delete-in-child-of-head.tentative_designMode=off_method=backspace-expected.txt

This is a testharness.js-based test.
[FAIL] delete around invisible <meta> should not delete it at joining paragraphs
  assert_in_array: The <div> elements should be merged without deleting <meta> value "<head><title>iframe</title><div>abcdef</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><meta></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><meta></head><body><br></body>", "<head><title>iframe</title><meta><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><meta><div>abcdef<br></div></head><body><br></body>"]
[FAIL] delete around invisible <title> should not delete it at joining paragraphs
  assert_in_array: The <div> elements should be merged without deleting <title> value "<head><title>iframe</title><div>abcdef</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><title></title></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><title></title></head><body><br></body>", "<head><title>iframe</title><title></title><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><title></title><div>abcdef<br></div></head><body><br></body>"]
[FAIL] delete around invisible <style> should not delete it at joining paragraphs
  assert_in_array: The <div> elements should be merged without deleting <style> value "<head><title>iframe</title><div>abcdef</div></head><body><br></body><style></style>" not in array ["<head><title>iframe</title><div>abcdef</div><style></style></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><style></style></head><body><br></body>", "<head><title>iframe</title><style></style><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><style></style><div>abcdef<br></div></head><body><br></body>"]
[FAIL] delete around invisible <script> should not delete it at joining paragraphs
  assert_in_array: The <div> elements should be merged without deleting <script> value "<head><title>iframe</title><div>abcdef</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><script></script></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><script></script></head><body><br></body>", "<head><title>iframe</title><script></script><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><script></script><div>abcdef<br></div></head><body><br></body>"]
[FAIL] delete around invisible <link> should not delete it at joining paragraphs
  assert_in_array: The <div> elements should be merged without deleting <link> value "<head><title>iframe</title><div>abcdef</div></head><body><br></body><link>" not in array ["<head><title>iframe</title><div>abcdef</div><link></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><link></head><body><br></body>", "<head><title>iframe</title><link><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><link><div>abcdef<br></div></head><body><br></body>"]
[FAIL] delete around invisible <base> should not delete it at joining paragraphs
  assert_in_array: The <div> elements should be merged without deleting <base> value "<head><title>iframe</title><div>abcdef</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><base></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><base></head><body><br></body>", "<head><title>iframe</title><base><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><base><div>abcdef<br></div></head><body><br></body>"]
[FAIL] delete around invisible <template> should not delete it at joining paragraphs
  assert_in_array: The <div> elements should be merged without deleting <template> value "<head><title>iframe</title><div>abcdef</div></head><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div><template></template></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div><template></template></head><body><br></body>", "<head><title>iframe</title><template></template><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><template></template><div>abcdef<br></div></head><body><br></body>"]
[FAIL] delete in visible <script> elements in <head> should not join them
  assert_equals: Visible <script> elements shouldn't be merged expected "<head><title>iframe</title><script>// abc</script><script>// def</script></head><body><br></body>" but got "<head><title>iframe</title><script>// abc// def</script></head><body><br></body>"
[FAIL] delete in visible <style> elements in <head> should not join them
  assert_equals: Visible <style> elements shouldn't be merged expected "<head><title>iframe</title><style>abc</style><style>def</style></head><body><br></body>" but got "<head><title>iframe</title><style>abcdef</style></head><body></body>"
[FAIL] delete in visible <title> elements in <head> should not join them
  assert_equals: Visible <title> elements shouldn't be merged expected "<head><title>iframe</title><title>abc</title><title>def</title></head><body><br></body>" but got "<head><title>iframe</title><title>abcdef</title></head><body><br></body>"
[FAIL] delete at boundary of <script> and <div> in <head> should not join them
  assert_equals: Visible <script> and <div> shouldn't be merged expected "<head><title>iframe</title><script>// abc</script><div>// def</div></head><body><br></body>" but got "<head><title>iframe</title><script>// abc// def</script></head><body><br></body>"
[FAIL] delete at boundary of <style> and <div> in <head> should not join them
  assert_equals: Visible <style> and <div> shouldn't be merged expected "<head><title>iframe</title><style>abc</style><div>def</div></head><body><br></body>" but got "<head><title>iframe</title><style>abcdef</style></head><body><br></body>"
[FAIL] delete at boundary of <title> and <div> in <head> should not join them
  assert_equals: Visible <title> and <div> shouldn't be merged expected "<head><title>iframe</title><title>abc</title><div>def</div></head><body><br></body>" but got "<head><title>iframe</title><title title=\\"display:block\\">abc</title><div>def</div></head><body><br></body>"
[FAIL] delete at boundary of <div> and <script> in <head> should not join them
  assert_equals: <div> and visible <script> shouldn't be merged expected "<head><title>iframe</title><div>// abc</div><script>// def</script></head><body><br></body>" but got "<head><title>iframe</title><div>// abc// def</div></head><body><br></body>"
[FAIL] delete at boundary of <div> and <style> in <head> should not join them
  assert_equals: <div> and visible <style> shouldn't be merged expected "<head><title>iframe</title><div>abc</div><style>def</style></head><body><br></body>" but got "<head><title>iframe</title><div>abcdef</div></head><body></body>"
[FAIL] delete at boundary of <div> and <title> in <head> should not join them
  assert_equals: <div> and visible <title> shouldn't be merged expected "<head><title>iframe</title><div>abc</div><title>def</title></head><body><br></body>" but got "<head><title>iframe</title><div>abcdef</div></head><body><br></body>"
Harness: the test ran to completion.
